requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Database and storage
minio>=7.2.0
redis>=5.0.0
chromadb>=0.4.0
neo4j>=5.0.0
weaviate-client>=3.25.0
qdrant-client>=1.7.0
python-arango>=7.5.0
elasticsearch==8.13.0
celery>=5.3.0
gevent>=23.0.0

# LLM and embeddings
openai>=1.0.0
anthropic>=0.7.0
llama-index>=0.9.40,<0.10.0
graphrag>=0.1.0
langfuse>=2.0.0
llama-cloud-services>=0.6.0

# Document processing
PyPDF2>=3.0.0
python-multipart>=0.0.6
llama-parser>=0.1.0
langchain-text-splitters>=0.0.1

# Utilities
httpx>=0.25.0
asyncio-mqtt>=0.16.0
nest_asyncio>=1.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
psycopg[binary,async]
